feat: provide functional sanoid spec file
This provides a functiona sanoid spec file which has been used in the
uCore project. This version of source is directly imported from the
universal blue project where I have been maintaining it for use in
Fedora CoreOS based systems.
https://github.com/ublue-os/packages/blob/02a988190cd5de730baca039c4c7d3d22d670c5b/ucore/sanoid.spec

Upstream does not build a package but they do provide a spec file. The
included spec file is modified version of upstream's which addressed
issues building on Fedora.
https://github.com/jimsalterjrs/sanoid/blob/538416879d45baa981307aeadd3abe34adae3874/packages/rhel/sanoid.spec